backend/requirements.txt
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
Faker==16.8.1
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
importlib-metadata==6.0.0
itsdangerous==2.1.2
MarkupSafe==2.1.1
pycparser==2.21
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==2.0.3
typing_extensions==4.4.0
Werkzeug==2.2.2
zipp==3.11.0
psycopg2-binary==2.9.9
yfinance==0.2.44
flask-cors==5.0.0
